refactor: Modular view actions
