+++ src/drive/web/modules/views/Recent/index.jsx
@@ -1,6 +1,11 @@
-import React, { useCallback } from 'react'
+import React, { useCallback, useContext } from 'react'
+
+import { SharingContext } from 'cozy-sharing'
+import { ModalContext } from 'drive/lib/ModalContext'
 
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
+import { useClient } from 'cozy-client'
+import { useDispatch } from 'react-redux'
 
 import FolderView from '../Folder/FolderView'
 import FolderViewHeader from '../Folder/FolderViewHeader'
@@ -9,6 +14,17 @@ import FolderViewBody from '../Folder/FolderViewBody'
 import { MobileAwareBreadcrumbV2 as Breadcrumb } from 'drive/web/modules/navigation/Breadcrumb/MobileAwareBreadcrumb'
 
 import useActions from 'drive/web/modules/actions/useActions'
+import {
+  share,
+  download,
+  trash,
+  open,
+  rename,
+  move,
+  qualify,
+  versions,
+  offline
+} from 'drive/web/modules/actions'
 
 import { buildRecentQuery } from 'drive/web/modules/queries'
 import { useFilesQueryWithPath } from './useFilesQueryWithPath'
@@ -26,7 +42,26 @@ export const RecentView = ({ router, children }) => {
     router.push(`/recent/file/${file.id}`)
   })
 
-  const actions = useActions({ hasWriteAccess: false, canMove: true })
+  const client = useClient()
+  const { pushModal, popModal } = useContext(ModalContext)
+  const { refresh } = useContext(SharingContext)
+  const dispatch = useDispatch()
+
+  const actionsOptions = {
+    client,
+    pushModal,
+    popModal,
+    refresh,
+    dispatch,
+    router,
+    location,
+    hasWriteAccess: false,
+    canMove: true
+  }
+  const actions = useActions(
+    [share, download, trash, open, rename, move, qualify, versions, offline],
+    actionsOptions
+  )
 
   return (
     <FolderView>
